// express/app/src/constants/errorCode.enum.ts
export enum EErrorCodes {
  // Common
  MISSING_FIELDS = 'MISSING_FIELDS',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  BAD_REQUEST = 'BAD_REQUEST',
  VALIDATION_ERROR = 'VALIDATION_ERROR',

  // Auth
  AUTH_LOGIN_FAIL = 'AUTH_LOGIN_FAIL',
  AUTH_UNAUTHORIZED = 'AUTH_UNAUTHORIZED',
  AUTH_FORBIDDEN = 'AUTH_FORBIDDEN',
  AUTH_TOKEN_EXPIRED = 'AUTH_TOKEN_EXPIRED',
  AUTH_MISSING_REFRESH_TOKEN = 'AUTH_MISSING_REFRESH_TOKEN',
  AUTH_INVALID_REFRESH_TOKEN = 'AUTH_INVALID_REFRESH_TOKEN',
  AUTH_LOGIN_METHOD_MISMATCH = 'AUTH_LOGIN_METHOD_MISMATCH',

  // Register
  AUTH_EMAIL_EXIST = 'AUTH_EMAIL_EXIST',
  AUTH_WRONG_PASSWORD = 'AUTH_WRONG_PASSWORD',
  AUTH_PASSWORD_NOT_SET = 'AUTH_PASSWORD_NOT_SET',
  AUTH_PASSWORD_ALREADY_SET = 'AUTH_PASSWORD_ALREADY_SET',

  // User
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
  USER_INACTIVE = 'USER_INACTIVE',

  // Resource / Business
  RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
  RESOURCE_CONFLICT = 'RESOURCE_CONFLICT',
  PRODUCT_OUT_OF_STOCK = 'PRODUCT_OUT_OF_STOCK',
  INSUFFICIENT_PERMISSION = 'INSUFFICIENT_PERMISSION',
}

// Tạo type để TypeScript gợi ý code
export type TErrorCode = keyof typeof EErrorCodes;